pyproject.toml
[project]
name = "vector-embedding-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "chardet>=5.2.0",
    "langchain>=0.3.25",
    "openai>=1.82.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "sentence-transformers>=4.1.0",
    "yoyo-migrations>=9.0.0",
]


[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib"
]
pythonpath = [
  "."
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]
